requirements.txt
django==2.1.1
django-webpack-loader
djangorestframework==3.9.0
django-rest-knox==3.4.0
geographiclib==1.49
googlemaps==3.0.2
bleach==3.0.2